Refactor V1

+++ eru/app/core/models.py
@@ -13,9 +13,9 @@ class ImageField(models.ImageField):
     def save_form_data(self, instance, data):
         if data is not None: 
             file = getattr(instance, self.attname)
-            if file != data:
+            if file and file != data:
                 file.delete(save=False)
-        super(ImageField, self).save_form_data(instance, data)
+        super().save_form_data(instance, data)
 
 class FileField(models.FileField):
     class Meta:
